move: InputTheme들 컴포넌트로 분리
